move authentication logic to `auth.rs`
